Fix footer capturing mouse events
If the browser window is too small, some mouse events will not be
registered by the game. This is because footer element takes up more
horizontal space that it needs. Changing footer width from "100%" to
"auto" will make it occupy only the required ammount of space, and not
capture mouse events intended for the bottom of game canvas.